consolidate classes
