+++ Src/MoneyFox/ViewModels/Statistics/StatisticCategorySpreadingViewModel.cs
@@ -3,7 +3,7 @@ using Microcharts;
 using MoneyFox.Application.Common;
 using MoneyFox.Application.Common.Interfaces;
 using MoneyFox.Application.Statistics;
-using MoneyFox.Application.Statistics.Queries.GetCategorySpreading;
+using MoneyFox.Application.Statistics.Queries;
 using MoneyFox.Ui.Shared.ViewModels.Statistics;
 using SkiaSharp;
 using System.Collections.ObjectModel;
@@ -53,13 +53,13 @@ namespace MoneyFox.ViewModels.Statistics
             }));
 
             var microChartItems = statisticItems.Select(x => new ChartEntry((float)x.Value)
-                                                {
-                                                    Label = x.Label,
-                                                    ValueLabel = x.ValueLabel,
-                                                    Color = SKColor.Parse(x.Color),
-                                                    ValueLabelColor = SKColor.Parse(x.Color)
-                                                })
-                                                .ToList();
+            {
+                Label = x.Label,
+                ValueLabel = x.ValueLabel,
+                Color = SKColor.Parse(x.Color),
+                ValueLabelColor = SKColor.Parse(x.Color)
+            })
+            .ToList();
 
             Chart = new DonutChart
             {
